Change operator result type from double to int
The idea for the result of the operation to be of double type was to
unify the result for all operators, including division, which returns
non-integer results. However, since the division operator implementation
is actually integer division, the result will also always be an int,
making the floating point precision of double useless.