Refractor code to force operator input first
I found it more logical to first input the operator itself, before its
operands.

An astute observer might pick up on the fact, that the changing of
sscanf call order in commandline mode does not actually do anything from
the perspective of the user and was purely made to make the code more
readable.